requirements.txt
Flask-JSON>=0.3.5
waitress>=2.1.2
PyYAML>=0.2.5
boto3>=1.26.66
twilio>=7.16.3